implemented guest wallet and chargable rooms

+++ src/main/java/people/Guest.java
@@ -18,7 +18,7 @@ public class Guest {
         return this.wallet;
     }
 
-    public void moneyLeavesWallet(int amount){
+    public void moneyLeavesWallet(double amount){
         this.wallet -= amount;
     }
